+++ src/plonemeeting/restapi/serializer/meeting.py
@@ -35,6 +35,7 @@ class SerializeMeetingToJsonBase(object):
         return result
 
     def _include_custom(self, obj, result):
+        """Include "date" by default."""
         if self.fullobjects or \
            "date" in self.metadata_fields or \
            self.get_param('include_base_data', True):