+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 Version 1.x is for PloneMeeting 4.1.x, version 2.x is for PloneMeeting 4.2.x+
 
 
-2.0 (unreleased)
+2.0 (2023-03-06)
 ----------------
 
 - Dropped support for `PloneMeeting 4.1.x`.